requirements.txt
asgiref==3.10.0
Django==5.2.8
django-ckeditor==6.7.3
django-js-asset==3.1.2
pillow==12.0.0
sqlparse==0.5.3
whitenoise==6.11.0

# Deployment dependencies
gunicorn==23.0.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
python-decouple==3.8
